refactor(engine): move reflection implementation into OperationDefinitionCompiler
Addresses review P2: the planning package now OWNS operation compilation,
not just a thin orchestration wrapper. All reflection/metadata work
(Kotlin reflection lookup, suspend detection, parameter-name discovery,
@System/@User processing, return-kind/type resolution, annotation
validation, shadowing warning) moved from OperationDefinition.create() in
TramaiEngine.kt into OperationDefinitionCompiler.compileDefinition().

OperationDefinition.create() is now a public compatibility facade
delegating to the compiler — API dump unchanged (apiCheck clean).

Also threads the precompiled plan fingerprint into runtime cache-key
construction (buildCacheKey accepts an optional operationFingerprint;
executeRaw/executeStructured pass plan.fingerprint instead of recomputing
it per invocation).
